V34 feat(ads): implement google mobile ads integration with rewarded and banner ads
- Add google_mobile_ads dependency and configure platform settings
- Create AdsService for managing ad initialization and loading
- Implement rewarded ads with bonus rewards and daily limits
- Add banner ads on mining press screen with toggle option
- Add admin dashboard for managing ad configurations
- Update session duration to support fractional hours

+++ lib/services/mining_state_service.dart
@@ -19,7 +19,7 @@ class MiningStateService extends ChangeNotifier {
   Timestamp? _lastStart;
   Timestamp? _lastEnd;
   bool _miningActive = false;
-  int _sessionHours = 24;
+  double _sessionHours = 24.0;
   int _streakDays = 0;
 
   // Manager state
@@ -58,7 +58,7 @@ class MiningStateService extends ChangeNotifier {
   int get managerMaxCommunity => _managerMaxCommunity;
   String? get activeManagerId => _activeManagerId;
   List<String> get managedCoinSelections => _managedCoinSelections;
-  int get sessionHours => _sessionHours;
+  double get sessionHours => _sessionHours;
   Timestamp? get lastEnd => _lastEnd;
   Timestamp? get lastStart => _lastStart;
   int get streakDays => _streakDays;
@@ -150,8 +150,9 @@ class MiningStateService extends ChangeNotifier {
         .get();
     final g = general.data() ?? {};
     _sessionHours =
-        ((g[FirestoreAppConfigFields.sessionDurationHours] as num?)?.toInt() ??
-        24);
+        ((g[FirestoreAppConfigFields.sessionDurationHours] as num?)
+                ?.toDouble() ??
+            24.0);
 
     // Load User Data
     final snap = await FirebaseFirestore.instance
